crm/requirements.txt
amqp==5.1.1
anyio==3.7.0
asgiref==3.7.2
Babel==2.12.1
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.2.7
dadata==21.10.1
Django==4.1.4
django-autocomplete-light==3.9.7
django-debug-toolbar==4.1.0
django-filter==23.2
django-hijack==3.4.1
django-model-utils==4.3.1
django-money==3.1.0
django-phonenumber-field==7.1.0
django-querysetsequence==0.17
flower==2.0.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
humanize==4.7.0
idna==3.4
kombu==5.3.1
phonenumbers==8.13.15
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
py-moneyed==2.0
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sentry-sdk==1.26.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tornado==6.3.2
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
